feat : [상품] 상품상세조회 시 랭킹정보 포함

+++ apps/commerce-api/src/main/java/com/loopers/domain/rank/Rank.java
@@ -13,10 +13,10 @@ import java.time.format.DateTimeFormatter;
 @Builder
 public class Rank {
     private final Long productId;
-    private final int position;
+    private final Long position;
     private final double score;
 
-    public static Rank create(Long productId, int position, double score) {
+    public static Rank create(Long productId, Long position, double score) {
         return new Rank(productId, position, score);
     }
 }
